+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Support for non-standard calendar model.
+
 ## [0.14.0] - 2023-09-22
 
 ### Changed